Feat: return the sum of two input numbers
Take input from HTML form, parse it in the backend and return the sum as
part of the sentence. The intent is to explain to the reader that
information taken in through a form can be processed beyond just simple
string parsing; addition is the easiest way to explain such a concept.