feat: updated template to explain Jinja better
The original content in the app.py file and the jinja_to_html.html file
was more or less a rip off of one of my other projects. Updated it to a
more consistent approach, and connected it to the previous blog content.
Now, we pass a packed dictionary containg name, choice, reason and sum
of the two numbers entered. The HTML is setup so as to render the keys
first and then generate a (hopefully) meaningful sentence out of it.